requirements.txt
streamlit
pandas
numpy
openai
openpyxl
python-dotenv